refactor: drop prefix from path
