+++ frontend/app/routes/sign-out/sign-out-route.tsx
